Allow arbitrarily nested Maybes on parameters and columns
I'm not entirely happy with this, but it seems to still make it through type
inference, and I can't think of a reasonable PostgreSQL representation for
Maybe Nothing to cause ambiguity.  It does make writing generic decoding
wrappers a lot easier.  (No use case for the parameters, yet, but for
consistency.)