Improve Sync/ReadyForQuery handling for #11
Keep better track of when we're expecting a ReadyForQuery, which
includes after a SimpleQuery and when a connection completes.  This
obviates the need to send a Sync, and eliminates the possibility of a
Sync and ReadyForQuery response crossing on the wire.  Previously, this
would trigger an unexpected second ReadyForQuery message that was
handled incorrectly.